refactored to use Recipient class
