+++ web/angular-wallet/src/app/layout/components/burst-recipient-input/burst-recipient-input.component.ts
@@ -15,6 +15,19 @@ export enum RecipientValidationStatus {
   VALID = 'valid',
 }
 
+export class Recipient {
+
+  constructor(
+    public addressRaw = '',
+    public amountNQT = '',
+    public addressRS = '',
+    public status = RecipientValidationStatus.UNKNOWN,
+    public type = RecipientType.UNKNOWN,
+  ) {
+
+  }
+}
+
 
 @Component({
   selector: 'burst-recipient-input',
@@ -22,18 +35,10 @@ export enum RecipientValidationStatus {
   styleUrls: ['./burst-recipient-input.component.scss']
 })
 export class BurstRecipientInputComponent implements OnInit {
-  // TODO: refactor to recipient class
-  recipientTypeValidationStatus = RecipientValidationStatus.UNKNOWN;
-  recipientType = RecipientType.UNKNOWN;
-  recipientValue = '';
-  recipientRS = '';
 
+  recipient = new Recipient();
   @Output()
   recipientChange = new EventEmitter();
-
-  set recipient(r: string) {
-    this.recipientValue = r;
-  }
 
   constructor(private accountService: AccountService) {
   }
@@ -42,25 +47,25 @@ export class BurstRecipientInputComponent implements OnInit {
   }
 
   applyRecipientType(recipient: string): void {
-    this.recipientRS = '';
-    this.recipientTypeValidationStatus = RecipientValidationStatus.UNKNOWN;
+    this.recipient.addressRS = '';
+    this.recipient.status = RecipientValidationStatus.UNKNOWN;
     const r = recipient.trim();
     if (r.length === 0) {
-      this.recipientType = RecipientType.UNKNOWN;
+      this.recipient.type = RecipientType.UNKNOWN;
     } else if (r.startsWith('BURST-')) {
-      this.recipientType = RecipientType.ADDRESS;
+      this.recipient.type = RecipientType.ADDRESS;
     } else if (/^\d+$/.test(r)) {
-      this.recipientType = RecipientType.ID;
+      this.recipient.type = RecipientType.ID;
     } else {
-      this.recipientType = RecipientType.ALIAS;
+      this.recipient.type = RecipientType.ALIAS;
     }
   }
 
 
   validateRecipient(): void {
     let accountFetchFn;
-    let id = this.recipientValue.trim();
-    switch (this.recipientType) {
+    let id = this.recipient.addressRaw.trim();
+    switch (this.recipient.type) {
       case RecipientType.ALIAS:
         accountFetchFn = this.accountService.getAlias;
         break;
@@ -79,29 +84,31 @@ export class BurstRecipientInputComponent implements OnInit {
     }
 
     accountFetchFn.call(this.accountService, id).then(({accountRS}) => {
-      this.recipientRS = accountRS;
-      this.recipientTypeValidationStatus = RecipientValidationStatus.VALID;
+      this.recipient.addressRS = accountRS;
+      this.recipient.status = RecipientValidationStatus.VALID;
     }).catch(() => {
-      this.recipientRS = '';
-      this.recipientTypeValidationStatus = RecipientValidationStatus.INVALID;
+      this.recipient.addressRS = '';
+      this.recipient.status = RecipientValidationStatus.INVALID;
     }).finally(() => {
         // TODO: use Recipient class
-        this.recipientChange.emit({
-          status: this.recipientTypeValidationStatus,
-          accountRS: this.recipientRS,
-          accountRaw: this.recipientValue,
-          type: this.recipientType
-        });
+        this.recipientChange.emit(this.recipient);
+
+        // this.recipientChange.emit({
+        //   status: this.recipientTypeValidationStatus,
+        //   accountRS: this.recipientRS,
+        //   accountRaw: this.recipientValue,
+        //   type: this.recipientType
+        // });
       }
     );
 
   }
 
-  getRecipientTypeName = (): string => RecipientType[this.recipientType];
+  getRecipientTypeName = (): string => RecipientType[this.recipient.type];
 
   getValidationHint(): string {
     // TODO: localization
-    switch (this.recipientTypeValidationStatus) {
+    switch (this.recipient.status) {
       case RecipientValidationStatus.UNKNOWN:
         return 'The address was not validated yet';
       case RecipientValidationStatus.VALID:
@@ -112,7 +119,7 @@ export class BurstRecipientInputComponent implements OnInit {
   }
 
   getValidationIcon(): string {
-    switch (this.recipientTypeValidationStatus) {
+    switch (this.recipient.status) {
       case RecipientValidationStatus.UNKNOWN:
         return 'help_outline';
       case RecipientValidationStatus.VALID:
@@ -123,6 +130,6 @@ export class BurstRecipientInputComponent implements OnInit {
   }
 
   getValidationClass(): string {
-    return 'badge ' + this.recipientTypeValidationStatus.toString().toLocaleLowerCase();
+    return 'badge ' + this.recipient.status.toString().toLocaleLowerCase();
   }
 }
