+++ web/angular-wallet/src/app/main/send-burst/send-burst-form/send-burst-form.component.ts
@@ -6,10 +6,12 @@ import {TransactionService} from 'app/main/transactions/transaction.service';
 import {NotifierService} from 'angular-notifier';
 import {I18nService} from 'app/layout/components/i18n/i18n.service';
 import {AccountService} from '../../../setup/account/account.service';
-import {Recipient} from '../typings';
 import {MatDialog, MatDialogRef} from '@angular/material';
 import {WarnSendDialogComponent} from '../warn-send-dialog/warn-send-dialog.component';
-import {RecipientValidationStatus} from '../../../layout/components/burst-recipient-input/burst-recipient-input.component';
+import {
+  Recipient,
+  RecipientValidationStatus
+} from '../../../layout/components/burst-recipient-input/burst-recipient-input.component';
 
 
 const isNotEmpty = (value: string) => value && value.length > 0;
@@ -126,10 +128,7 @@ export class SendBurstFormComponent implements OnInit {
       isNotEmpty(this.pin);
   }
 
-  onRecipientChange(recipient: any): void {
-    this.recipient.addressRS = recipient.accountRS;
-    this.recipient.status = recipient.status;
-    this.recipient.addressRaw = recipient.accountRaw;
-    this.recipient.type = recipient.type;
+  onRecipientChange(recipient: Recipient): void {
+    this.recipient = recipient;
   }
 }
